Added channelNames & chats
